STYLE: Change to presentational about detailBook
- Change to presentational about detailBook
- Added header search event
