+++ src/Components/Routes.js
@@ -22,11 +22,11 @@ const LoggedInRoutes = () => (
     <Switch>
       <Route exact path="/" component={Main} />
       <Route exact path="/myPage" component={MyPage} />
-      <Route exact path="/search" component={Search} />
       <Route exact path="/seeAllPost" component={SeeAllPost} />
       <Route exact path="/seeBestPost" component={SeeBestPost} />
       <Route exact path="/seeMyPost" component={SeeMyPost} />
       <Route exact path="/writePost" component={WritePost} />
+      <Route exact path="/search" component={Search} />
       <Route exact path="/book-list/:gener" component={BookList} />
       <Route exact path="/:itemId" component={BookDetail} />
       <Redirect from="*" to="/" />
