+++ src/Routes/BookDetail/index.js
@@ -1,0 +1,3 @@
+import BookDetailContainer from "./BookDetailContainer";
+
+export default BookDetailContainer;
